backend/app/requirements.txt
# ==============================================================================
# Python Dependencies
# ==============================================================================
# This file is organized into sections for clarity.
# Use `pip install -r requirements.txt` to install all dependencies.
# Versions are aligned with google-adk's official pyproject.toml for compatibility.
# ==============================================================================

# ------------------------------------------------------------------------------
# Application Framework
# ------------------------------------------------------------------------------
fastapi>=0.115.0
uvicorn>=0.34.0
gunicorn==22.0.0
asgiref>=3.6.0

# ------------------------------------------------------------------------------
# Google Cloud & Firebase Services
# ------------------------------------------------------------------------------
# Google Agent Development Kit
google-adk>=1.3.0

# Google Cloud & ADK Core Dependencies
firebase-admin==6.9.0
firebase-functions>=0.2.0
google-cloud-aiplatform[agent_engines]>=1.95.1
google-cloud-firestore==2.19.0
google-cloud-secret-manager>=2.22.0
google-cloud-speech>=2.30.0
google-cloud-storage>=2.18.0,<3.0.0
google-genai>=1.17.0

# Google API client libraries
google-api-python-client>=2.157.0
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# ------------------------------------------------------------------------------
# PDF, Image & HTML Processing
# ------------------------------------------------------------------------------
weasyprint==62.3
Pillow==10.4.0
jinja2>=3.1.2
beautifulsoup4==4.12.3
# Optional: for better Japanese font support
fonttools>=4.43.0

# ------------------------------------------------------------------------------
# Utilities
# ------------------------------------------------------------------------------
# HTTP clients
requests>=2.32.4
aiohttp>=3.9.0
# Environment variable management
pydantic-settings>=1.2.0
python-dotenv==1.0.1
litellm>=1.63.11

# ------------------------------------------------------------------------------
# Development & Testing
# ------------------------------------------------------------------------------
pytest==8.2.2
pytest-mock==3.14.0
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0